safari-534.53-branch/LayoutTests/http/tests/security/dataURL/xss-DENIED-from-javascript-url-window-open-expected.txt

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8000/security/dataURL/xss-DENIED-from-javascript-url-window-open.html from frame with URL data:text/html,%3Chtml%3E%3Chead%3E%3Cscript%3Efunction%20test()%20{try%20{opener.document.getElementById(%22accessMe%22).innerHTML%20=%20%22FAIL:%20Access%20from%20a%20window%20opened%20with%20a%20data:%20URL%20was%20allowed!%22;}%20catch%20(e)%20{}if%20(window.layoutTestController)layoutTestController.globalFlag%20=%20true;}%3C/script%3E%3C/head%3E%3Cbody%20onload=%22test();%22%3E%3Cp%3EOpened%20Frame.%3C/p%3E%3C/body%3E%3C/html%3E. Domains, protocols and ports must match.

Opener Frame

PASS: Access from a window opened with a data: URL was denied.
